♻️ 2A32BY blueprints: decompose catalog parsers
