+++ .agentplane/tasks/202605282346-2A32BY/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605282346-2A32BY`
+Title: Blueprint catalog parser decomposition
+Canonical task record: `.agentplane/tasks/202605282346-2A32BY/README.md`
+
+## Summary
+
+Blueprint catalog parser decomposition
+
+Decompose packages/agentplane/src/commands/blueprints/catalog.ts by extracting catalog/manifest types and parsers while preserving catalog load/install behavior.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/blueprints/catalog.ts by extracting catalog/manifest types and parsers while preserving catalog load/install behavior.
+- Out of scope: unrelated refactors not required for "Blueprint catalog parser decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T23:46:36.926Z
+- Branch: task/202605282346-2A32BY/blueprint-catalog-parser-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
