+++ .agentplane/tasks/202605282346-2A32BY/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 2A32BY task: Blueprint catalog parser decomposition [202605282346-2A32BY]
